"1/ ui 100% ( display list of towns from local + screen to add town + deatil weather) 2/ get data from local when offline 3/ optimization extensions + ToolbarView + add loadingState"

+++ lib/src/main/java/com/example/lib/data/remote/dataSource/WeatherRemote.kt
@@ -13,8 +13,8 @@ class WeatherRemote @Inject constructor() {
     lateinit var serviceEndPoint: ServiceEndPoint
 
     suspend fun getWeatherTownAsync(
-        lat: Long,
-        lon: Long,
+        lat: Double,
+        lon: Double,
         exclude: String,
         appId: String,
         units: String,
