+++ lib/src/main/java/com/example/lib/domain/RetrieveWeatherInfoFromLocal.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
  * Created by zaineb on 26/07/2020
  */
 class RetrieveWeatherInfoFromLocal @Inject constructor(private val weatherRepository: WeatherRepository) {
-    suspend fun execute(lat: Long, lon: Long) =
+    suspend fun execute(lat: Double, lon: Double) =
         weatherRepository.getWeatherInfoFromLocal(lat, lon)
 }
